3-print_all.c

+++ variadic_functions/3-print_all.c
@@ -32,7 +32,7 @@ void print_all(const char * const format, ...)
 	break;
 
 	case 'f':
-	printf("%f", va_arg(args, float));
+	printf("%f", va_arg(args, double));
 	break;
 
 	case 's':
